experiment/cryptorex-exp/CryptoREX/requirements.txt
angr==9.2.13
python_magic==0.4.27
rarfile==4.0
